Fix warning in memory-hwm

+++ profiling/memory-hwm/kp_hwm.cpp
@@ -39,7 +39,7 @@ void kokkosp_init_library(const int loadSeq, const uint64_t interfaceVer,
   printf(
       "KokkosP: High Water Mark Library Initialized (sequence is %d, version: "
       "%llu)\n",
-      loadSeq, interfaceVer);
+      loadSeq, (unsigned long long)(interfaceVer));
 }
 
 // darwin report rusage.ru_maxrss in bytes
